bcrypt: In unixlib, don't use Windows long type.

+++ dlls/bcrypt/bcrypt_internal.h
@@ -155,7 +155,7 @@ struct algorithm
     struct object hdr;
     enum alg_id   id;
     enum mode_id  mode;
-    ULONG         flags;
+    unsigned      flags;
 };
 
 struct key_symmetric
@@ -165,7 +165,7 @@ struct key_symmetric
     UCHAR       *vector;
     ULONG        vector_len;
     UCHAR       *secret;
-    ULONG        secret_len;
+    unsigned     secret_len;
     CRITICAL_SECTION cs;
 };
 
@@ -174,9 +174,9 @@ struct key_symmetric
 struct key_asymmetric
 {
     ULONG             bitlen;     /* ignored for ECC keys */
-    ULONG             flags;
+    unsigned          flags;
     UCHAR            *pubkey;
-    ULONG             pubkey_len;
+    unsigned          pubkey_len;
     DSSSEED           dss_seed;
 };
 
@@ -208,7 +208,7 @@ struct key_symmetric_encrypt_params
 {
     struct key  *key;
     const UCHAR *input;
-    ULONG        input_len;
+    unsigned     input_len;
     UCHAR       *output;
     ULONG        output_len;
 };
@@ -217,7 +217,7 @@ struct key_symmetric_decrypt_params
 {
     struct key  *key;
     const UCHAR *input;
-    ULONG        input_len;
+    unsigned     input_len;
     UCHAR       *output;
     ULONG        output_len;
 };
@@ -233,7 +233,7 @@ struct key_asymmetric_decrypt_params
 {
     struct key  *key;
     UCHAR       *input;
-    ULONG        input_len;
+    unsigned     input_len;
     UCHAR       *output;
     ULONG        output_len;
     ULONG       *ret_len;
@@ -250,11 +250,11 @@ struct key_asymmetric_sign_params
     struct key  *key;
     void        *padding;
     UCHAR       *input;
-    ULONG        input_len;
+    unsigned     input_len;
     UCHAR       *output;
     ULONG        output_len;
     ULONG       *ret_len;
-    ULONG        flags;
+    unsigned     flags;
 };
 
 struct key_asymmetric_verify_params
@@ -262,10 +262,10 @@ struct key_asymmetric_verify_params
     struct key *key;
     void       *padding;
     UCHAR      *hash;
-    ULONG       hash_len;
+    unsigned    hash_len;
     UCHAR      *signature;
     ULONG       signature_len;
-    ULONG       flags;
+    unsigned    flags;
 };
 
 struct key_export_params
